% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{addText}
\alias{addText}
\title{Generates HTML text paragraphs}
\usage{
addText(..., clear = T, style = "")
}
\value{
none, just pre-allocate in the main workspace
}
\description{
Generates HTML text paragraphs
}
\examples{
addText("<h1 style="...">Heading</h1>","<p>Content</p>", list(bullets, style=c("...","...")), style="<for div>")
}
