% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{aloneNumber}
\alias{aloneNumber}
\title{Detection of alone numbers preceded/followed by NAs}
\usage{
aloneNumber(x)
}
\arguments{
\item{string}{input character vector}
}
\value{
true/false if an orphan found.
}
\description{
Detection of alone numbers preceded/followed by NAs
}
\examples{
aloneNumber(c(1,2,3,NA,99,NA))
}
