% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{download2fix}
\alias{download2fix}
\title{Update functionality locally}
\usage{
download2fix(
  repoClone = "C:/jaac/WORK/R utils/plotly/plotlyjs",
  localPlayground = "C:/jaac/WORK/R utils/plotly"
)
}
\value{
none
}
\description{
Update functionality locally
}
