% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{figure}
\alias{figure}
\title{Compilation of HTML/plotly document}
\usage{
figure(
  ...,
  x = c(),
  type = c(),
  title = "",
  vline = c(),
  legend = c(),
  xlabel = "",
  ylabel = "",
  colors = c(),
  by = "",
  lineWidth = 2,
  reportFile = "tmp.html",
  libFile = "path/to/file/plotly.min.js",
  lightWeight = F,
  reopen = F
)
}
\value{
none, modifies the given HTML template file.
}
\description{
Compilation of HTML/plotly document
}
\examples{
figure(c(1:5))
}
