% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{fileparts}
\alias{fileparts}
\title{Folder, file and extension parts of a file name.}
\usage{
fileparts(filepath)
}
\arguments{
\item{string}{input character vector}
}
\value{
list of file parts.
}
\description{
Folder, file and extension parts of a file name.
}
\examples{
fileparts("filePath")
}
