% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{graphCustom}
\alias{graphCustom}
\title{Custom graph}
\usage{
graphCustom(
  grObj,
  objInfo,
  igr,
  objs,
  nobjs,
  x,
  title,
  vline,
  legend,
  xlabel,
  ylabel,
  colors,
  by,
  lineWidth,
  layout,
  config,
  width,
  height
)
}
\value{
none
}
\description{
Custom graph
}
