% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{graphTypeBar}
\alias{graphTypeBar}
\title{Bar graph}
\usage{
graphTypeBar(
  grObj,
  objInfo,
  igr,
  objs,
  nobjs,
  x,
  title,
  legend,
  xlabel,
  ylabel,
  colors,
  width,
  height
)
}
\value{
none
}
\description{
Bar graph
}
