% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{graphTypeLine}
\alias{graphTypeLine}
\title{Line graph}
\usage{
graphTypeLine(
  grObj,
  objInfo,
  igr,
  objs,
  nobjs,
  x,
  title,
  vline,
  legend,
  xlabel,
  ylabel,
  colors,
  lineWidth,
  width,
  height
)
}
\value{
none
}
\description{
Line graph
}
