% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{graphTypeTS}
\alias{graphTypeTS}
\title{Time series graph}
\usage{
graphTypeTS(grObj, grInfo, igr, objs, nobjs, x, title, vline, legend,
  xlabel, ylabel, colors, lineWidth)
}
\value{
none
}
\description{
Time series graph
}
