% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{plotlyIni}
\alias{plotlyIni}
\title{Initializer for plotly graphs (deletes all preceding graphs)}
\usage{
plotlyIni()
}
\value{
none, just pre-allocate in the main workspace
}
\description{
Initializer for plotly graphs (deletes all preceding graphs)
}
\examples{
plotlyIni()
}
