% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{processOpts}
\alias{processOpts}
\title{list() -> JSON generator (mainly for graph trace defs)}
\usage{
processOpts(input, level)
}
\value{
JSON structure
}
\description{
list() -> JSON generator (mainly for graph trace defs)
}
