% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{push2github}
\alias{push2github}
\title{Update functionality locally (export not needed for this fcn!)}
\usage{
push2github(repoClone)
}
\value{
none
}
\description{
Update functionality locally (export not needed for this fcn!)
}
