% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{uploadFixed}
\alias{uploadFixed}
\title{Update functionality locally (export not needed for this fcn!)}
\usage{
uploadFixed(
  repoClone = "C:/jaac/WORK/R utils/plotly/plotlyjs",
  localPlayground = "C:/jaac/WORK/R utils/plotly"
)
}
\value{
none
}
\description{
Update functionality locally (export not needed for this fcn!)
}
