% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{addGraph}
\alias{addGraph}
\title{Generates HTML/plotly code for a new graph}
\usage{
addGraph(..., x = c(), type = "line", title = "", vline = c(),
  legend = c(), xlabel = "", ylabel = "", colors = c(), size = 6,
  lineWidth = 2)
}
\value{
none, just pre-allocate in the main workspace
}
\description{
Generates HTML/plotly code for a new graph
}
\examples{
addGraph(list(periods, values))
addGraph(list(periods, values),options)
}
