% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{addGraph}
\alias{addGraph}
\title{Generates HTML/plotly code for a new graph}
\usage{
addGraph(
  ...,
  x = c(),
  type = c(),
  title = "",
  vline = c(),
  legend = c(),
  xlabel = "",
  ylabel = "",
  colors = c(),
  by = "",
  lineWidth = NA,
  width = 600,
  height = 450,
  style = "",
  clear = F,
  layout = list(),
  config = list()
)
}
\value{
none, just pre-allocate in the main workspace
}
\description{
Generates HTML/plotly code for a new graph
}
\examples{
addGraph(list(periods, values))
addGraph(list(periods, values),options)
}
