% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{figure}
\alias{figure}
\title{Compilation of HTML/plotly document}
\usage{
figure(d, reportFile = "tmp.html", reopen = F)
}
\value{
none, modifies the given HTML template file.
}
\description{
Compilation of HTML/plotly document
}
\examples{
figure(c(1:5))
}
