% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{figure}
\alias{figure}
\title{Compilation of HTML/plotly document}
\usage{
figure(
  ...,
  x = c(),
  type = c(),
  title = "",
  vline = c(),
  legend = c(),
  xlabel = "",
  ylabel = "",
  colors = c(),
  by = "",
  lineWidth = 2,
  width = 600,
  height = 450,
  layout = list(),
  config = list(),
  style = "",
  reportFile = "tmp.html",
  libFile = "path/to/file/plotly.min.js",
  lightWeight = F,
  name = "",
  css = "",
  font = "",
  reopen = F,
  debug = F
)
}
\value{
none, modifies the given HTML template file.
}
\description{
Compilation of HTML/plotly document
}
\examples{
figure(c(1:5))
}
