% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlyGraphs.R
\name{graphTypeBubble}
\alias{graphTypeBubble}
\title{Bubble graph}
\usage{
graphTypeBubble(
  grObj,
  objInfo,
  igr,
  objs,
  nobjs,
  title,
  legend,
  xlabel,
  ylabel,
  colors,
  by,
  width,
  height
)
}
\value{
none
}
\description{
Bubble graph
}
